requirements.txt
rdkit
pubchempy
requests
openai
pandas
python-dotenv